Rework talkfile creation window.
Support selecting multiple folders by replacing the provided path with a tree
view that allows selecting multiple entries. The view is limited to the
selected mountpoint (i.e. the player) so this removes the possibility of
generating talk files for files that are not on the player. However, since
Rockbox Utility disables most functionality without an attached player this
isn't too much of a problem. Creating a standalone application for talkfile
creation that is not limited to the player is possible as well.

Change-Id: Ic68e7556f2e2e5b9c121aaba759a42a4d1d9d53a