Support release-candidate entry format for releases.
The release-candiate entries provide both version and URL instead of only a
version entry and requiring the URL to be constructed separately. Since it's
easy to identify the format support both for release entries for improved
flexibility -- if the server URLs ever change the server only needs to get
updated to use the extended format.

Change-Id: If5c669b72d7acd6ac648c5f42dd844ae36904df6