Update patch set 2
Patch Set 2: (2 inline comments)

I have not checked if there is conflict with any other part of rockbox, but the playlist viewer was using (potentially) all of the plugin buffer before this change too.

First there is the playlist index buffer that got allocated the first half of the plugin buffer when it's needed.

Then the second half gets allocated to the name buffer.

Now before my change not all of this memory was actually used. It was in fact limited by MAX_PLAYLIST_ENTRIES * MAX_PATH which was about 52000 bytes.

I read somewhere that on small targets the plugin buffer is 32 kilobytes, so in that case rockbox was already using the full plugin buffer. On 512kilobyte targets 205 kilobytes of plugin buffer are left unused (in the worst case), but the code doesn't make this clear anywhere.

Please see some inline comments that i also made

Patch-set: 2
Label: Code-Review=0