Update patch set 4
Patch Set 4:

I really think not doing more heuristics and defaulting to notouch is the better way (also for maintainability of the code). Your already idea doesn't if there's a progressbar and volume bar. What's wrong with defaulting to notouch? I think such a change in behavior where it's not clear how existing themes handle should not be on by default.

Patch-set: 4