Implement basic glyph fallback.
Currently, if a font is missing a unicode character is usually displays
a placeholder "box" character instead. This change makes it so that the
font code automatically looks for suitable fonts to load and provide the
missing glyph.

When a missing glyph is requested, a table (hardcoded for now, will be
loaded from a config file later) of fonts is searched. Each font of a
matching size is loaded and checked for presence of the required glyph.
When a match is found that font is used to draw that specific glyph
instead.

Minor changes to the .fnt file format were required to be able to
identify missing glyphs in a font, which were previously just redirected
to a dummy character, but now have an unlikely value in the width table.

Change-Id: I861fc4d1abd061848a3e22ec858eb49501a4f379