Implement basic glyph fallback.
Currently, if a font is missing a unicode character is usually displays
a placeholder "box" character instead. This change makes it so that the
font code automatically looks for suitable fonts to load and provide the
missing glyph.

When a missing glyph is requested, a table (hardcoded for now, will be
loaded from a config file later) of fonts and respective ranges of
glyphs which they may provide is searched. When a match is found the
font is loaded and the code attempts to use it to provide the missing
glyph.

The missing glyph detection code is curently very simple and misses
missing glyphs inside the font's valid range of characters. Some
heuristic algorithm will need to be added to cover these cases.

Change-Id: I861fc4d1abd061848a3e22ec858eb49501a4f379