Update patch set 8
Patch Set 8:

If the buffer is full it branches out at the top of the function. It does print a message if anything is dropped. Do you think a count is really needed? I imagine a couple KB won't be exceeded in normal use and debug builds can use much larger buffers if needed.

Patch-set: 8