Introduce new hermite polynomial resampler.
Currently, the linear resampler is left in as a compile time option.
Eventually we may want to remove it entirely depending on how fast the hermite
resampler is on older targets as the quality is much better using hermite.

Based on work in g#231 by Michael Sevakis and I, but remove real networks
code in favor of my own implementation.

Change-Id: I145d2169dcbe70d5011691e4109d4c8ff1e95434