scrobbler: fix double scrobbling in some cases
  The current algorithm of scrobbler shouldn't scrobble
tracks which have been already played for more than
a half of their length when we are resuming playback,
cause it obviously creates double scrobbling of the same track.
  But it wasn't working correctly because PLAYBACK_EVENT_TRACK_CHANGE
for a first few times was being sent when codec hasn't been started yet,
so track's elapsed time at that moment equals 0.
  I've just made custom function for a starting of scrobbler which
is called in audio_pcmbuf_sync_position which I believe is the first
time when elapsed time for a resuming track is set.
  From looks of me, it fixes the problem.

Change-Id: Ib9f739b3d59ccf7ec64d1366c4a1f647685a5dc4