Update patch set 10
Patch Set 10:

One more thing I realised while reviewing the patch, is multichannel support. In general, we should decide what to when we encounter a multichannel opus file. I propose we just skip it for now.

There is still a FIXME in the code for properly skipping over a table in the opus headers in case of multichannel files. In other words, the current opus parsing code might crash on scanning a multichannel opus.

Patch-set: 10
Label: Code-Review=0
Label: Tested=0