ELF loader for plugins
This is early snapshot of my work.

Currently it is possible to build plugins
as relocatable elfs and run most.

special cases spotted so far:
* mpegplayer doesnt run on ARM as it links
libmad-mpeg.a which is built without -mlong-calls
*doom doesn't load as symtab is too big.
*chessbox data aborts (don't know why)
*text_editor shows black screen but you can quit
*vu_meter stops refreshing screen after ~2s

The loader is based on code from prex 0.9
but heavily tweaked. Currently it reads
relocs directly form ELF but symtab is
cached in memory at the end of pluginbuf.
The downside is that doom stopped loading.

SH and MIPS binutils are too old to support
--gc-sections and -r at the same time. Building
without --gc-sections leads to sagnificant bin
size increase. Moreover stip on MIPS is broken
and leaves unneeded content.

Tested only on PP only.

Changelog:
* cache symtab in memory
  support COMMON symbols
  fixes in plugin_elf.lds and plugins.make
  make utils/elfloader usefull
* added strip to the plugins.make (thanks Frank)
* innitial commit

Change-Id: I0334add810e5e0c96196df452c3ab0b83b655685