ELF loader for plugins
This is snapshot of my work.

Currently it is possible to build plugins
as relocatable elfs and run most.

special cases spotted so far:
* mpegplayer   - black screen, backlight reacts,
                 can't quit
* chessbox     - data aborts (don't know why)
* text_editor  - it doesn't quite work but
                 you can quit
* vu_meter     - stops refreshing screen after ~2s
* oscilloscope - no signal
* rockboy      - data aborts or undef instr
* fft          - hangs, backlight reacts

The loader is based on code from prex 0.9
but heavily tweaked. Currently it reads
relocs directly form ELF but symtab is
cached in memory at the end of pluginbuf.

SH and MIPS binutils are too old to support
--gc-sections and -r at the same time. Building
without --gc-sections leads to sagnificant bin
size increase. Moreover stip on MIPS is broken
and leaves unneeded content.

Tested on PP only.

Changelog:
* V4
- defer .bss and COMMON zeroing so memory
  load time requirements are relaxed
- proof-of-concept veneer insertion on ARM
- drop -mlong-calls switch from PLUGINFLAGS

* V3
- cache symtab in memory
- support COMMON symbols
- fixes in plugin_elf.lds and plugins.make
- make utils/elfloader usefull

* V2
- added strip to the plugins.make (thanks Frank)

* V1
- initial commit

Change-Id: I0334add810e5e0c96196df452c3ab0b83b655685