ELF loader for plugins
This is snapshot of my work.

Currently it is possible to build plugins
as relocatable elfs and run most.

special cases spotted so far:
* mpegplayer   - black screen, backlight reacts,
                 can't quit (tested on PP only)
* chessbox     - data aborts (don't know why PP)
* text_editor  - it doesn't quite work but
                 you can quit (PP)
* vu_meter     - stops refreshing screen after ~2s
                 (PP, n2g, CF)
* oscilloscope - no signal (PP, n2g, CF)
* rockboy      - data aborts or undef instr
                 (tested on PP only)
* fft          - hangs, backlight reacts (PP)
                 or report no playback (CF, n2g)

The loader is based on code from prex 0.9
but heavily tweaked. Currently it reads
relocs directly form ELF but symtab is
cached in memory at the end of pluginbuf.

SH and MIPS binutils are too old to support
--gc-sections and -r at the same time. Building
without --gc-sections leads to sagnificant bin
size increase. Moreover stip on MIPS is broken
and leaves unneeded content.

Tested on Vibe500(PP502x), n2g and
MPIO HD300(MCF5249)

Changelog:
* V5
- support RELA needed by Coldfire
- fixes in plugin_elf.lds
- PC side test tool can handle big endian
  ELFs now as well

* V4
- defer .bss and COMMON zeroing so memory
  load time requirements are relaxed
- proof-of-concept veneer insertion on ARM
- drop -mlong-calls switch from PLUGINFLAGS

* V3
- cache symtab in memory
- support COMMON symbols
- fixes in plugin_elf.lds and plugins.make
- make utils/elfloader usefull

* V2
- added strip to the plugins.make (thanks Frank)

* V1
- initial commit

Change-Id: I0334add810e5e0c96196df452c3ab0b83b655685