ELF loader for plugins
This is early snapshot of my work.

Currently it is possible to build plugins
as relocatable elfs and run some.

The loader is based on code from prex 0.9
but heavily tweaked. Currently it reads
relocs and symbols info directly from ELF
to not use additional memory for file loading.
This have the downside that it is somewhat slow
for bigger plugins. Will look at this later.

SH and MIPS binutils are too old to support
--gc-sections and -r at the same time. Building
without --gc-sections leads to sagnificant bin
size increase. Moreover stip on MIPS is broken
and leaves unneeded content.

Tested only on PP only.

Changelog:
* added strip to the plugins.make (thanks Frank)
* innitial commit

Change-Id: I0334add810e5e0c96196df452c3ab0b83b655685