Update patch set 1
Patch Set 1:

I was trying (and failing) to copy the main font twice since the second time it used the path instead of the filename. This patch fixes that and copies both fonts and sets the correct paths for both in the theme .cfg
Are you sure you read the patch in the right direction?

Patch-set: 1
Reviewer: Gerrit User 1000015 <1000015@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0