Create change
Patch-set: 1
Change-id: I31ab41c4120f4af64eb6998b7e7b6f9051585efb
Subject: Add microSD automatic mount and unmount to Samsung YP-R0. The patch adds a thread in order to check whether the memory card is inserted or not, and if so it handles the correct system calls. I tested it a little, bugs can still be present. Moreover is it really a good way just sleeping before mounting the card? To notice also the possibility to get stuck into the filebrowser when removing the card while playing/browsing...
Branch: refs/heads/master
Commit: c2578fd0eee9152c8c56707b6b5b6f2e28011548