skin_engine: User selectable track data display
Adds 2 new tags to the skin engine:
%iu and %Iu which cause the skin code from a different skin
to be drawn in the current viewport. this allows the user
to control what data they want displayed and not be stuck
with what the themer wants.

a new skin file named metadata.txt should exist in the
wps directory with a number of viewports (specially named).
The viewports should not do any drawing other than plain text
tags (prefereably limited to %i* and %I* but this is not yet
enforced).

Viewports should be named like this:
this_X - where X is the number of text lines this viewport
should be used to draw..

i.e
%Vl(this_2, -, -, -, -, -)
%ia %it
%s%in

Viewports for the next tracks data are named next_X (X being
the line count).

When %iu (or %Iu) is found, metadata.txt is searched for a viewport
matching the amount of lines which could fit in the viewport with %iu.
i.e this_6, then this_5 then this_4 until one is found.

Themers are requested to replace all the %i* lines with a %iu.

Changelog:
V1 - initial working version
v2 - Made it work better if it shares the viewport with other tags.
     i.e %v(...)
         %iu

         %Iu
     will now work finding the best combination to fit in the viewport.

Change-Id: Ibfacd1c83b127bf93d2339a2820742b7128637f1