Update patch set 4
Patch Set 4:

I agree with wodz. I would prefer to not have this in our codebase.

With this you still need to know a subset of the skin language and look at a page where all the rest of the language is also explained. So, to me this is a very awkward feature with very little gain, but a lot of added complexity.

Patch-set: 4
Label: Code-Review=0
Label: Tested=0