Update patch set 2
Patch Set 2:

There are no more hardcoded glyph ranges to search, just a list of fonts. (Which I think should be moved to some configuration file? I'm unsure what's the preferred way of offering these kinds of configs.)

That means I had to change the font format to signal which glyphs are present in the font when searching. Also means several chained searches need to be done in the font caches, which is inefficient but which I don't see a way around.

Possible idea would be to implant the replacement glyphs into the original font's cache, but that has the issue of it maybe no fitting into the space allocated for glyphs on each entry due to having a different maxwidth.

Patch-set: 2
Label: Code-Review=0