Added the screenshots for the Sansa Clip Zip manual.
Missing screenshots:

Invadrox, which I can't find in the menu, but on IRC they said it's
for LCD only, so it should be removed.

Pacbox, Rockboy and Zxbox: it's unclear where to find the ROMs.

Poor quality image:
manual/rockbox_interface/images/sansaclipzip-front.png. It can work,
but my “artistic” skills are not enough. Anyway, the information on
the picture is correct.

All the screenshots were taken using the default theme and the
simulator compiled for the Sansa Clip Zip.

Change-Id: I779a3b8fbf9a1ab2f12b8f8e6bcf55f2363900ce