Fix .apk signs when using OpenJDK7 (Android builds)
This fix the "jarsigner error: java.security.SignatureException: private
key algorithm is not compatible with signature algorithm" when compiling
Android builds using OpenJDK7. This patch explicit how keytools make a
debug key so it is compatible with the changes in jarsigner made on
OpenJDK7. It is still working with OpenJDK6.

For those who already have a debug key, it will be necessary to remove
the previous key located on ~/.android.

Change-Id: Ifa4d37d0bda76e8937dc18572a7e8586c1a18a63
Reviewed-on: http://gerrit.rockbox.org/359
Reviewed-by: Frank Gevaerts <frank@gevaerts.be>