Allow comparison of multiple tags with same value in DB search
Add the ability to search multiple tags with the same input string, by adding a clause that refers to the value of the previous clause.
The clause is called "#prev#" and is added to the list of 'special meaning' clauses which are evaluated during runtime.
This allows adding queries in the tagnavi file to search multiple tags, for instance:
A query that searches for the same keyword (that's given via input keyboard) within the album and the title tags:
"Album or title"-> album ? album ~ "" | title ~ "#prev#" -> title = "fmt_title"
The "#prev#" keyword can be concatenated more than one times, to search within more than 2 tags. It also supports numeric values.

Change-Id: Ic4e985538c9ac8f0cf4e8b31fabb945899cfed4d