rk27xx: usb driver
This driver works for me but needs further cleanup and testing.
Many thanks to pamaury for massive help.

According to simple benchmark I got ~2.38 MB/s when copying
big file over usb to sd card. This is comparable to the
transfer rate of the rb sd driver itself.
OF is about 2x slower !!! More tests are needed though.

Change-Id: Iee3036944652fd6431d3177ab619e5df1f9bd44c