Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Is it allowed to examine these variables ourselves to not lose the ability to observe the scan in the UI and interrupt it? Or, perhaps apply the hardware tuning incrementally and read the "tuned" output on the chip?

Patch-set: 1
Reviewer: Gerrit User 1000041 <1000041@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=-1
Label: Tested=0