Update patch set 6
Patch Set 6:

Marcin: Sorry, You're completely missing the point of the feature.

The aim is:
1) to make it easier for Joe User to modify the track info he sees on his DAP without needing to understand Random Skinner's theme

2) To have a consistent set of information displayed *that the user wants* (as opposed to the themer).

Think of it like this...
I make a theme with only 3 lines of text for the track, my files are all tagged as normal and I listen to non compilation cds (i.e most music cds), but someone wants to use my theme who only listens to podcasts or classic music which is tagged differently (iirc the composer tag is more important than the artist in classic tracks) so they will want to always show different info than I would want.
With this patch all they need to do (if the skin i set up correctly) is modify trackinfo.txt and whatever theme they use the info will be the same.

This feature wont work well with themes which use icons and bitmaps to display the text, but most just set aside one (or a few) text areas which this would work very well with.

From V5 I've modified it so even if a theme has 2 lines at the top and 2 lines at the bottom (With AA in the middle) it can still be setup to display 2 different sets of information.

Patch-set: 6
Reviewer: Gerrit User 1000011 <1000011@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0