Update patch set 1
Patch Set 1:

I didn't see it useful for non-touchscreen, but I'm not opposed to including it on all.

I probably need to redo the patch, because it's broken when the line scrolls (in which case the scroll engine draws over the separator). Doing it likewise to gradient line selector is probably the way to go.

Patch-set: 1
Label: Code-Review=0
Label: Tested=0