Update patch set 1
Patch Set 1:

In reponse to IRC: Yes, swap the actual data or switch to out-of-place configuration when using this setting and output a copy of the buffer but with swapped pointers. After thinking about it, I can't see any problem arising with the patch in combination with that is _currently_ in the tree but it does break the in-place expectation of only processing samples but not modifying the buffer structure itself.

Patch-set: 1