Update patch set 2
Patch Set 2:

I think it would be terribly confusing to do that.

And if the user tried to edit one of the defaults without first reverting to the "Flat" preset the frequency stepping would be off for whatever 5 of the 10 bands you chose to use.

That, and, I don't see a reason not to.
Actually, primarily that. I would find it very as a new user to find that the default presets only used 50% of the bands available.

Patch-set: 2
Label: Code-Review=0