Update patch set 1
Patch Set 1:

It can, but that doesn't necessarily coordinate all channels and fading on stop/pause also gets more complicated. Also, it won't affect direct PCM users that don't use the mixer (or the DSP). It really should go in the PCM driver just like with the sim so that all PCM is scaled no matter how its used.

Patch-set: 1