Implement universal in-PCM-driver software volume control.
Implements double-buffered volume, balance and prescaling control in
the main PCM driver when HAVE_SW_VOLUME_CONTROL is defined ensuring
that all PCM is volume controlled. Supports -73 to +6 dB using a 15-bit
factor so that no large-integer math is needed.

Low-level hardware drivers do not have to implement it themselves but
parameters can be changed (currently defined in pcm-internal.h) to work
best with a particular SoC or to provide different volume ranges.

Volume and prescale calls should be made in the codec driver. It should
appear as a normal hardware interface.

NOTE: Have some junk Gigabeat F/X, iRiver and x5  edits for testings
purposes and so this particular patch shouldn't be pushed. Will remove
once tested.

Change-Id: Idf6316a64ef4fb8abcede10707e1e6c6d01d57db